+++ tests/Feature/Service/Student/SpecializationListServiceTest.php
@@ -6,7 +6,7 @@ namespace Tests\Feature\Service\Student;
 
 use Tests\TestCase;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
-use App\Service\SpecializationListService;
+use App\Service\Student\SpecializationListService;
 use Database\Factories\ResumeFactory;
 use App\Models\Resume;
